Fix bug when scaling canvas with CSS
Previously, setting the canvas width and height via CSS "width" and "height" properties
would create a scale multiplier that SignaturePad did not account for. This caused the
line to show up offset from where the user attempted to draw.  

Now, the scale multiplier is taken into account, and scaling the canvas with CSS works
just fine, with lines appearing directly under the cursor.